// web/src/client/types.gen.ts
export type EnvVar = {
  /**
   * Name
   */
  name: string;
  /**
   * Value
   */
  value: string;
};
// ...
export type TaskConfigListItem = {
  /**
   * Name
   */
  name: string;
  /**
   * Description
   */
  description?: string | null;
  /**
   * Env Vars
   */
  env_vars?: Array<EnvVar> | null;
  /**
   * Id
   */
  id: number;
  /**
   * Project Id
   */
  project_id: number;
  /**
   * Created At
   */
  created_at: Date;
  /**
   * Updated At
   */
  updated_at: Date;
  owner: UserSearch;
};
// ...
export type UserSearch = {
  /**
   * Id
   */
  id: number;
  /**
   * Email
   */
  email: string;
  /**
   * Name
   */
  name: string;
  /**
   * En Name
   */
  en_name?: string | null;
  /**
   * Union Id
   */
  union_id?: string | null;
  /**
   * Avatar
   */
  avatar?: string | null;
};